feat(utilities/utilities): move breakpoints array and type to shared file | sas
