+++ packages/utilities/projects/utilities/src/js/mediaQuery/breakpoint.ts
@@ -1,3 +1,4 @@
+import type { Breakpoint } from './breakpointShared';
 import { breakpointBase } from './breakpointBase';
 import { breakpointXS } from './breakpointXS';
 import { breakpointS } from './breakpointS';
@@ -5,9 +6,6 @@ import { breakpointM } from './breakpointM';
 import { breakpointL } from './breakpointL';
 import { breakpointXL } from './breakpointXL';
 import { breakpointXXL } from './breakpointXXL';
-
-export const breakpoints = ['base', 'xs', 's', 'm', 'l', 'xl', 'xxl'] as const;
-export type Breakpoint = (typeof breakpoints)[number];
 
 export const breakpoint: { [key in Breakpoint]: number } = {
   base: breakpointBase,
