+++ packages/utilities/projects/utilities/src/js/mediaQuery/breakpointShared.ts
@@ -1,0 +1,2 @@
+export const breakpoints = ['base', 'xs', 's', 'm', 'l', 'xl', 'xxl'] as const;
+export type Breakpoint = (typeof breakpoints)[number];
