+++ packages/utilities/projects/utilities/src/js/mediaQuery/getMediaQueryMin.ts
@@ -1,5 +1,5 @@
 import type { BKey } from './mediaQueryShared';
-import type { Breakpoint } from './breakpoint';
+import type { Breakpoint } from './breakpointShared';
 import { breakpoint } from './breakpoint';
 
 export function getMediaQueryMin(min: BKey<'base'>): '@media(min-width:0px)';
