+++ packages/utilities/projects/utilities/src/js/mediaQuery/mediaQueryShared.ts
@@ -1,3 +1,3 @@
-import type { Breakpoint } from './breakpoint';
+import type { Breakpoint } from './breakpointShared';
 
 export type BKey<K extends Breakpoint> = Extract<Breakpoint, K>;
